requirements.txt
asgiref==3.8.1
Django==5.1.6
psycopg2-binary==2.9.10
sqlparse==0.5.3
django-embed-video
environs[django]==9.5.0

